refact: rename header to basicHeader
